Move scanner code into BooksTable
